split functionality into different files
