+++ solai_evolutionary_algorithm/evolution/evolver_config.py
@@ -1,0 +1,15 @@
+from dataclasses import dataclass
+from typing import List, Optional, Tuple, Callable
+
+from solai_evolutionary_algorithm.evolution.evolution_types import InitialPopulationProducer, FitnessEvaluation, \
+    PopulationEvolver, EndCriteria, Population, EvaluatedPopulation
+from solai_evolutionary_algorithm.evolution.evolver_listener import EvolverListener
+
+
+@dataclass(frozen=True)
+class EvolverConfig:
+    initial_population_producer: InitialPopulationProducer
+    fitness_evaluator: FitnessEvaluation
+    population_evolver: PopulationEvolver
+    end_criteria: EndCriteria
+    evolver_listeners: Optional[List[EvolverListener]] = None
