+++ solai_evolutionary_algorithm/evolution/evolver_listener.py
@@ -1,0 +1,11 @@
+
+class EvolverListener:
+
+    def on_start(self, config):
+        pass
+
+    def on_new_generation(self, new_popoulation, is_last_generation):
+        pass
+
+    def on_end(self):
+        pass
